[SPARK-14842][SQL] Implement view creation in sql/core
