+++ sql/catalyst/src/main/scala/org/apache/spark/sql/types/UserDefinedType.scala
@@ -87,6 +87,8 @@ abstract class UserDefinedType[UserType >: Null] extends DataType with Serializa
     case that: UserDefinedType[_] => this.acceptsType(that)
     case _ => false
   }
+
+  override def simpleString: String = sqlType.simpleString
 }
 
 /**
